backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
requests>=2.32.0
beautifulsoup4>=4.12.3
lxml>=5.3.0
playwright>=1.49.0
httpx>=0.28.0
python-dotenv>=1.0.1
sqlalchemy>=2.0.36
aiosqlite>=0.20.0
selenium>=4.27.0
webdriver-manager>=4.0.2
apscheduler>=3.10.4
cryptography>=42.0.0
